refactor (phase 3)
